MINOR: Move ByteBuffer streams to internal package
Generated-by: Codex
