+++ clients/src/main/java/org/apache/kafka/common/compress/Compression.java
@@ -17,8 +17,8 @@
 package org.apache.kafka.common.compress;
 
 import org.apache.kafka.common.record.internal.CompressionType;
-import org.apache.kafka.common.utils.ByteBufferOutputStream;
 import org.apache.kafka.common.utils.internals.BufferSupplier;
+import org.apache.kafka.common.utils.internals.ByteBufferOutputStream;
 
 import java.io.InputStream;
 import java.io.OutputStream;
