+++ clients/src/main/java/org/apache/kafka/common/compress/GzipCompression.java
@@ -18,9 +18,9 @@ package org.apache.kafka.common.compress;
 
 import org.apache.kafka.common.KafkaException;
 import org.apache.kafka.common.record.internal.CompressionType;
-import org.apache.kafka.common.utils.ByteBufferInputStream;
-import org.apache.kafka.common.utils.ByteBufferOutputStream;
 import org.apache.kafka.common.utils.internals.BufferSupplier;
+import org.apache.kafka.common.utils.internals.ByteBufferInputStream;
+import org.apache.kafka.common.utils.internals.ByteBufferOutputStream;
 import org.apache.kafka.common.utils.internals.ChunkedBytesStream;
 
 import java.io.BufferedOutputStream;
