+++ clients/src/main/java/org/apache/kafka/common/compress/Lz4Compression.java
@@ -19,8 +19,8 @@ package org.apache.kafka.common.compress;
 import org.apache.kafka.common.KafkaException;
 import org.apache.kafka.common.record.internal.CompressionType;
 import org.apache.kafka.common.record.internal.RecordBatch;
-import org.apache.kafka.common.utils.ByteBufferOutputStream;
 import org.apache.kafka.common.utils.internals.BufferSupplier;
+import org.apache.kafka.common.utils.internals.ByteBufferOutputStream;
 import org.apache.kafka.common.utils.internals.ChunkedBytesStream;
 
 import java.io.InputStream;
