+++ clients/src/main/java/org/apache/kafka/common/record/internal/MemoryRecordsBuilder.java
@@ -26,8 +26,8 @@ import org.apache.kafka.common.message.SnapshotHeaderRecord;
 import org.apache.kafka.common.message.VotersRecord;
 import org.apache.kafka.common.protocol.MessageUtil;
 import org.apache.kafka.common.record.TimestampType;
-import org.apache.kafka.common.utils.ByteBufferOutputStream;
 import org.apache.kafka.common.utils.Utils;
+import org.apache.kafka.common.utils.internals.ByteBufferOutputStream;
 
 import java.io.DataOutputStream;
 import java.io.IOException;
