+++ clients/src/main/java/org/apache/kafka/common/record/internal/MutableRecordBatch.java
@@ -17,8 +17,8 @@
 package org.apache.kafka.common.record.internal;
 
 import org.apache.kafka.common.record.TimestampType;
-import org.apache.kafka.common.utils.ByteBufferOutputStream;
 import org.apache.kafka.common.utils.internals.BufferSupplier;
+import org.apache.kafka.common.utils.internals.ByteBufferOutputStream;
 import org.apache.kafka.common.utils.internals.CloseableIterator;
 
 /**
