+++ clients/src/test/java/org/apache/kafka/common/compress/GzipCompressionTest.java
@@ -19,8 +19,8 @@ package org.apache.kafka.common.compress;
 import org.apache.kafka.common.config.ConfigDef;
 import org.apache.kafka.common.config.ConfigException;
 import org.apache.kafka.common.record.internal.RecordBatch;
-import org.apache.kafka.common.utils.ByteBufferOutputStream;
 import org.apache.kafka.common.utils.internals.BufferSupplier;
+import org.apache.kafka.common.utils.internals.ByteBufferOutputStream;
 
 import org.junit.jupiter.api.Test;
