+++ clients/src/test/java/org/apache/kafka/common/utils/internals/ByteUtilsTest.java
@@ -16,8 +16,6 @@
  */
 package org.apache.kafka.common.utils.internals;
 
-import org.apache.kafka.common.utils.ByteBufferInputStream;
-import org.apache.kafka.common.utils.ByteBufferOutputStream;
 import org.apache.kafka.common.utils.Bytes;
 import org.apache.kafka.common.utils.Utils;
