+++ clients/src/test/java/org/apache/kafka/common/utils/internals/ChunkedBytesStreamTest.java
@@ -16,8 +16,6 @@
  */
 package org.apache.kafka.common.utils.internals;
 
-import org.apache.kafka.common.utils.ByteBufferInputStream;
-
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
